Adds functionality to get and change postcode reference based on user's country
This functionality allows for the retrieval of the correct terminology of postcode based on the country of the user.

USA -> zip code
Canada -> postal code
default -> postcode

This also updates any reference to postcode to the correct terminology when the user changes their country, i.e Canada to USA - postal code to zip code.

 🐿 v2.12.3