Add placeholer override attribute
JSX escapes all HTML attributes which causes issues for the
placeholder in the delivery instructions component. Because a line
break entity is used to put instructions across multiple lines this
breaks our tests as JSX will escape this. By adding the ability to
override the placeholder we can have two working but different default
placehodlers for JSX and Handlebars but still have tests pass.

 🐿 v2.12.4