Customise payment-term component for lite subs
A lite subscription gives access to the subset of the FT for a fraction of the cost of a regular sub. Due to the low cost of the subscription, we don’t want to subscription to auto-renew.

Amends the payment-term component to remove mentions of auto-renewal if the offer is fixed term (e.g. in the case of lite subscription).